boa.acf {boa}	R Documentation

Autocorrelation Function

Description

Computes lag autocorrelations for the parameters in an MCMC sequence.

Usage


boa.acf(link, lags)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link)

lags 	 Vector of lags at which to estimate the autocorrelation function.

Value

A matrix whose columns and rows contain the estimated autocorrelation functions at the specified lags and the monitored parameters, respectively.

Author(s)

Brian J. Smith

See Also

boa.plot, boa.plot.acf, boa.print.acf

[Package boa Index]