boa.batchMeans {boa}	R Documentation

Batch Means

Description

Computes batch means for the parameters in an MCMC sequence.

Usage


boa.batchMeans(link, size)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

size 	 Number of iterations to include in each batch.

Value

A matrix whose columns and rows contain the means within sequential batches and the monitored parameters, respectively.

Author(s)

Brian J. Smith

[Package boa Index]