boa.chain {boa}	R Documentation

Session Lists of MCMC Sequences

Description

Extracts and sets the MCMC sequence(s) stored internally for use during a BOA session. Users can safely use this function to extract data. However, it should not be called directly to modify the stored MCMC sequence(s). Instead, data management should accomplished with the libboa.chain functions.

Usage


boa.chain(...)

Arguments

... 	 A list may be given as the only argument, or a character string given as the only argument, or any number of arguments may be in the <name> = <value> form, or no argument at all may be given. See the Value and Side Effects sections for further details.

Value

If no arguments are supplied, a list of the current values is returned. If a character string is given, the current value of the named variable is returned. Otherwise, no values are returned.

Parameter Names and Default Values

master = list()

Master list containing all added MCMC sequences. Sequences may only be added to or deleted from the master list. The sequences in this list are never modified or used in any of the analyses. It serves as a template for the working list described below. The parameters and iterations may differ between the sequences. Each element in the list is a matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names are stored in the dimnames.

master.support = list()

List containing the support for each of the parameters in 'master'. Each element in the list is a matrix whose columns and rows contain the monitored parameters and the support (lower and upper limits), respectively. The default for each parameter is c(-Inf, Inf).

work = master

Working list used in all analyses. This list is essentially a modifiable copy of 'master'. All data management functions operate on the working list.

work.support = master.support

List containing the support for each of the parameters in 'work'.

work.sync = TRUE

Logical value indicating that the working list is an identical copy of the master list. This variable is automatically set to F after a successful call to boa.chain.subset or boa.chain.collapse. While work.sync = F, only the master list will be updated if MCMC sequences are added or deleted. A call to boa.chain.reset will reset this variable to T and copy the master list to the working list.

Side Effects

When variables are set, boa.chain() modifies the internal list .boa.chain. If boa.chain() is called with either a list as the single argument, or with one or more arguments in the <name> = <value> form, the variables specified by the names in the arguments are modified.

Author(s)

Brian J. Smith

[Package boa Index]