boa.chain.add {boa}	R Documentation

Add MCMC Sequence

Description

Adds an MCMC sequence to the session lists of sequences. Most users should not call this function directly. Passing improperly formatted data to this function will most likely break the functions in this library. The preferred method of adding a new sequence is to call the function boa.chain.import.

Usage


boa.chain.add(link, lname)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

lname 	 Character string giving the list name to use for the newly added MCMC sequence.

Value

A logical value indicating that 'link' has been successfully added.

Author(s)

Brian J. Smith

See Also

boa.chain.import

[Package boa Index]