boa.chain.collapse {boa}	R Documentation

Concatenate MCMC Sequences

Description

Concatenates together all of the MCMC sequences in the working session list of sequences. Only those parameters common to all sequences are kept.

Usage


boa.chain.collapse()

Value

A logical value indicating that the MCMC sequences have been collapsed. If the sequences share no common parameters, no modifications are made to the working list and FALSE is returned.

Author(s)

Brian J. Smith

See Also

code{boa.chain.reset}, code{boa.chain.subset}

[Package boa Index]