boa.chain.del {boa}	R Documentation

Delete MCMC Sequences

Description

Delete MCMC sequences from the session list of sequences.

Usage


boa.chain.del(lnames, pnames)

Arguments

lnames 	 Character vector giving the names of the MCMC sequences in the session list of sequences to be deleted. If omitted, no sequences are deleted.

pnames 	 Character vector giving the names of the parameters in the MCMC sequences to be deleted. If omitted, no parameters are deleted.

Side Effects

The specified MCMC sequences are deleted from the session lists of sequences.

Author(s)

Brian J. Smith

[Package boa Index]