boa.chain.eval {boa}	R Documentation

Formulate New Parameter

Description

Evaluates a user-specified expression to formulate a new parameter in each of the system lists of MCMC sequences.

Usage


boa.chain.eval(expr, pname)

Arguments

expr 	 S expression object used to formulate the new parameter. Typically, the new parameter is a function of the existing parameters.

pname 	 Name given to the new parameter.

Side Effects

The specified parameter is added to the session lists of MCMC sequences.

Author(s)

Brian J. Smith

[Package boa Index]