boa.chain.import {boa}	R Documentation

Import MCMC Sequence

Description

Reads an MCMC sequence from an external file and adds it to the session lists of sequences.

Usage


boa.chain.import(prefix, path = boa.par("path"), type = "ASCII")

Arguments

prefix 	 Character string giving the prefix for the file(s) in which the MCMC sequence is stored.

path 	 Character string giving the directory path in which the file(s) subsides.

type 	 Character string specifying the type of data to be imported. Currently, the supported types are:

"ASCII" : ASCII text file. Calls boa.importASCII to import data from "prefix.txt".
"BUGS" : BUGS output files. Calls boa.importBUGS to import data from "prefix.ind" and "prefix.out".
"S" : S data frame or matrix.

Value

A logical value indicating that the specified file(s) has been successfully imported.

Side Effects

The imported MCMC sequence is automatically added to the session lists of sequences via a call to boa.chain.add.

Author(s)

Brian J. Smith

See Also

boa.chain.add, boa.importASCII, boa.importBUGS

[Package boa Index]