boa.chain.info {boa}	R Documentation

Information about MCMC Sequences

Description

Returns summary information from a list of MCMC sequences.

Usage


boa.chain.info(chain, chain.support)

Arguments

chain 	 List of matrices whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to the dimnames.

chain.support 	 List of matrices whose columns and rows contain the monitored parameters and the support (lower and upper limits), respectively.

Value

lnames 	 Character vector giving the names of the MCMC sequences in the session list of sequences.

pnames 	 List of character vectors giving the parameter names in each of the MCMC sequences.

iter 	 List of numeric vectors giving the iterations from each MCMC sequence.

iter.range 	 Matrix whose columns give the range of the iterations for the MCMC sequences named in the rows.

support 	 List of numeric vectors giving the support for the parameters in each of the MCMC sequences.

Author(s)

Brian J. Smith

See Also

boa.print.info

[Package boa Index]