boa.chain.reset {boa}	R Documentation

Reset MCMC Sequences

Description

Returns the data in the working list of MCMC sequences to the state it was in when originally imported. This function undoes any subsetting or collapsing that was done via the boa.chain.collapse or boa.chain.subset functions.

Usage


boa.chain.reset()

Side Effects

Returns the data in the working list of MCMC sequences to the state it was in when originally imported. This function undoes any subsetting or collapsing that was done via the boa.chain.collapse or boa.chain.subset functions.

Author(s)

Brian J. Smith

See Also

boa.chain.collapse, boa.chain.subset

[Package boa Index]