boa.chain.subset {boa}	R Documentation

Subset MCMC Sequences

Description

Selects a subset of the MCMC sequences stored in the working session list of sequences.

Usage


boa.chain.subset(lnames, pnames, iter)

Arguments

lnames 	 Character vector giving the names of the MCMC sequences to include in the subset. If omitted, all sequences are included.

pnames 	 Character vector giving the names of the parameters to include in the subset. If omitted, all parameters are included.

iter 	 Character vector giving the names of the parameters to include in the subset. If omitted, all parameters are included.

Value

A logical value indicating that the data has been successfully subsetted. If the requested subset contains no data, the working session list is not modified and FALSE is returned.

Side Effects

The subsetted data, if not an empty set, is copied to the working session list of MCMC sequences used in all analyses.

Author(s)

Brian J. Smith

See Also

boa.chain.collapse, boa.chain.reset

[Package boa Index]