boa.getiter {boa}	R Documentation

Extract Iterations

Description

Extracts a subset of the iterations in an MCMC sequence.

Usage


boa.getiter(link, iter)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

iter 	 Numeric vector containing the iterations to extract from 'link'.

Value

A subset of 'link' containing all prior parameters and the iterations specified in 'iter'.

Author(s)

Brian J. Smith

See Also

boa.iter, boa.getparms

[Package boa Index]