boa.getparms {boa}	R Documentation

Extract Parameters

Description

Extracts a subset of the parameters in an MCMC sequence.

Usage


boa.getparms(link, pnames)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

pnames 	 Character vector containing the names of the parameters to extract from 'link'.

Value

A subset of 'link' containing the specified parameters in 'pnames' and all prior iterations.

Author(s)

Brian J. Smith

See Also

boa.getiter, boa.pnames

[Package boa Index]