boa.geweke {boa}	R Documentation

Geweke Convergence Diagnostics

Description

Computes the Geweke convergence diagnostics for the parameters in an MCMC sequence.

Usage


boa.geweke(link, p.first, p.last)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

p.first 	 Proportion of iterations to include in the first window.

p.last 	 Proportion of iterations to include in the last window.

Value

A matrix whose columns contain the Z-Scores and associated p-values and whose rows contain the monitored parameters.

Author(s)

Brian J. Smith

References

Geweke, J. (1992). Evaluating the accuracy of sampling-based approaches to calculating posterior moments. In Bayesian Statistics 4, (ed. J. M. Bernardo, J. O. Berger, A. P. Dawid, and A. F. M. Smith). Clarendon Press, Oxford, UK.

See Also

boa.plot, boa.plot.geweke, boa.print.geweke

[Package boa Index]