boa.gewekePwr {boa}	R Documentation

Geweke Spectral Density Estimation

Description

Estimates the spectrum frequency for the parameters in an MCMC sequence.

Usage


boa.gewekePwr(link)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

Value

A vector of spectrum frequencies for the monitored parameters.

Author(s)

Brian J. Smith

[Package boa Index]