boa.hpd {boa}	R Documentation

Highest Probability Density Interval

Description

Estimates the highest probability density (HPD) interval for the given parameter draws. Uses the Chen and Shao algorithm assuming a unimodal marginal posterior distribution.

Usage


boa.hpd(x, alpha)

Arguments

x 	 MCMC draws from the marginal posterior to use in computing the HPD.

alpha 	 Specifies the 100*(1 - alpha)% interal to compute.

Value

A vector containing the lower and upper bound of the HPD interval, labeled "Lower Bound" and "Upper Bound", respectively.

Author(s)

Brian J. Smith

References

Chen, M-H. and Shao, Q-M. (1999). Monte Carlo estimation of Bayesian credible and HPD intervals. Journal of Computational and Graphical Statistics, 8(1), 69-92.

[Package boa Index]