boa.importBUGS {boa}	R Documentation

Import BUGS Data

Description

Import data from BUGS output files.

Usage


boa.importBUGS(prefix, path = NULL)

Arguments

prefix 	 Character string giving the prefix for the files in which the BUGS output is stored. boa.importBUGS() looks for the two files "prefix.ind" and "prefix.out".

path 	 Character string giving the directory path in which the file subsides. This argument may be omitted if the file is located in the current working directory. The specified path should not end with a slash(es).

Value

If the data is successfully imported, a matrix is returned whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names are stored in the dimnames of the returned matrix. Otherwise, NULL is returned.

Author(s)

Brian J. Smith

See Also

boa.chain.import, boa.importASCII, boa.importMatrix

[Package boa Index]