boa.init {boa}	R Documentation

Start Session

Description

boa.init() is the first function called to start a BOA session when using the command-line. It sets up the internal data structures and initializes them with the appropriate default values.

Usage


boa.init(recover = FALSE)

Arguments

recover 	 a logical value for use if the previous BOA menu session terminated unexpectedly. It may be possible to crash the menu system by supplying the wrong type of data. In the event of a crash, recover = TRUE may be used to ensure that no data is lost.

Author(s)

Brian J. Smith

See Also

boa.load, boa.quit

[Package boa Index]