boa.load {boa}	R Documentation

Load Session

Description

Loads the data and global parameter settings from a previously saved BOA session.

Usage


boa.load(name, envir = globalenv())

Arguments

name 	 Character string giving the name of the object containing the session to be loaded.

envir 	 The 'environment' from which the object should come. For more information, consult the help documentation in R on the get function.

Value

A logical value indicating that the object was successfully loaded.

Author(s)

Brian J. Smith

[Package boa Index]