boa.menu {boa}	R Documentation

Interactive BOA Menu

Description

Starts the interactive menu for BOA. Provides a menu-driven interface to all of the convergence diagnostics and statistical and graphical analysis tools.

Usage


boa.menu(recover = FALSE)

Arguments

recover 	 a logical value for use if the previous BOA menu session terminated unexpectedly. It may be possible to crash the menu system by supplying the wrong type of data. In the event of a crash, recover = TRUE may be used to ensure that no data is lost.

Warning

Do not use the recover argument unless starting the menu immediately after a crash.

Note

boa.menu is the only function in the BOA library that need be called in order to use the menu-driven interface.

Author(s)

Brian J. Smith

[Package boa Index]