boa.plot {boa}	R Documentation

Plot Parameters in MCMC Sequences

Description

Automatically generates the specified plot type for all parameters in the working session list of MCMC sequences. This function takes care of all the plotting tasks - opening windows, setting the number of plots per page, and adding titles.

Usage


boa.plot(type, dev = boa.par("dev"), mfdim = boa.par("plot.mfdim"),
newplot = boa.par("plot.new"), onelink = boa.par("plot.onelink"),
title = boa.par("title"))

Arguments

type 	 The type of plots to generate. The supported types are:

"acf" = autocorrelation functions
"bandg" = Brooks and Gelman multivariate shrink factors
"density" = density functions
"gandr" = Gelman and Rubin shrink factors
"geweke" = Geweke convergence diagnostics
"history" = running means
"trace" = trace histories

dev 	 Character string giving the name of the function that creates graphics windows on the current platform. For Unix systems this is either "motif", "openlook", or "X11". For windows, this should be set to "win.graph"

mfdim 	 Numeric vector giving the maximum number of rows and columns, respectively, of plots to include in a single graphics window. If only one graphics window is opened, 'mfdim' is proportionately scaled down so as to minimize the number of empty frames within that window.

newplot 	 Logical value indicating that a new graphics window be automatically opened. Otherwise, previous graphics windows will be closed.

onelink 	 Logical value indicating that each plot should include only one MCMC sequence. Otherwise, all sequences are displayed on the same plot.

title 	 Logical value indicating that a title be added to the plot.

Value

A logical value indicating that the plots were successfully created.

Author(s)

Brian J. Smith

See Also

boa.plot.acf, boa.plot.bandg, boa.plot.density, boa.plot.gandr, boa.plot.geweke, boa.plot.history, boa.plot.trace

[Package boa Index]