boa.plot.acf {boa}	R Documentation

Plot Autocorrelation Function

Description

Creates a single plot of the lag autocorrelations for a specified parameter.

Usage


boa.plot.acf(lname, pname, annotate = boa.par("legend"))

Arguments

lname 	 Character string giving the name of the desired MCMC sequence in the working session list of sequences.

pname 	 Character string giving the name of the parameter to be plotted.

annotate 	 Logical value indicating that a legend be included in the plot.

Value

A logical value indicating that the plot was successfully created.

Author(s)

Brian J. Smith

See Also

boa.acf, boa.print.acf

[Package boa Index]