boa.plot.close {boa}	R Documentation

Close Graphics Windows

Description

Closes graphics windows opened with boa.plot.open().

Usage


boa.plot.close(which = dev.cur())

Arguments

which 	 Integer vector specifying the graphics windows to be closed.

Value

If the specified graphics windows are successfully closed, the number of the current graphics window is returned. Otherwise NULL is returned.

Author(s)

Brian J. Smith

See Also

boa.plot.open

[Package boa Index]