boa.plot.density {boa}	R Documentation

Plot Density Functions

Description

Estimates and displays, in a single plot, the density function(s) for the specified parameter(s).

Usage


boa.plot.density(lnames, pname, bandwidth = boa.par("bandwidth"),
window = boa.par("kernel"), annotate = boa.par("legend"))

Arguments

lnames 	 Character vector giving the names of the desired MCMC sequence in the working session list of sequences.

pname 	 Character string giving the name of the parameter to be plotted.

bandwidth 	 Function for computing the bandwidth used in estimating the density functions for parameters. This should take one argument which is a numeric vector of data on which density estimation is to be performed. A constant bandwidth may be specified by having this function return the desired constant.

window 	 Character string giving the type of window used in estimating the density functions for the parameters. Available choices are "cosine", "gaussian", "rectangular", or "triangular".

annotate 	 Logical value indicating that a legend be included in the plot.

Value

A logical value indicating that the plot was successfully created.

Author(s)

Brian J. Smith

See Also

boa.plot

[Package boa Index]