boa.plot.geweke {boa}	R Documentation

Plot Geweke Convergence Diagnostics

Description

Creates a single plot of the Geweke convergence diagnostics within different line segments for a parameter in an MCMC sequence.

Usage


boa.plot.geweke(lname, pname, bins = boa.par("geweke.bins"), p.first = boa.par("geweke.first"), p.last = boa.par("geweke.last"), alpha = boa.par("alpha"), annotate = boa.par("legend"))

Arguments

lname 	 Character string giving the name of the desired MCMC sequence in the working session list of sequences.

pname 	 Character string giving the name of the parameter to be plotted.

bins 	 Number of line segments within the MCMC sequence at which to plot the Geweke convergence diagnostics. The ith line segment contains the last ((bins - i + 1) / bins)*100 bins. This may lead to segments in which there are too few iterations to compute the convergence diagnostics. Such segments, if they exist, are automatically omitted from the calculations.

p.first 	 Proportion of iterations to include in the first window.

p.last 	 Proportion of iterations to include in the last window.

alpha 	 Alpha level for the rejection region lines drawn on the graph.

annotate 	 Logical value indicating that a legend be included in the plot.

Value

A logical value indicating that the plot was successfully created.

Author(s)

Brian J. Smith, Nicky Best, Kate Cowles

See Also

boa.geweke, boa.plot, boa.print.geweke

[Package boa Index]