boa.plot.history {boa}	R Documentation

Plot Parameter Estimation Histories

Description

Computes and displays, in a single plot, the running mean(s) for the specified parameter(s).

Usage


boa.plot.history(lnames, pname, annotate = boa.par("legend"))

Arguments

lnames 	 Character vector giving the name of the desired MCMC sequence in the working session list of sequences.

pname 	 Character vector giving the names of the parameters to be plotted.

annotate 	 Logical value indicating that a legend be included in the plot.

Value

A logical value indicating that the plot was successfully created.

Author(s)

Brian J. Smith

See Also

boa.plot

[Package boa Index]