boa.plot.par {boa}	R Documentation

Set Plotting Parameters

Description

Set Plotting Parameters

Usage


boa.plot.par(mfdim = c(1, 1), title = TRUE)

Arguments

mfdim 	 Numeric vector with two elements giving the number of rows and column, respectively, of plots to display on the page. Plots will be drawn row-by-row.

title 	 Logical value indicating that a title be added to the plot.

Author(s)

Brian J. Smith

[Package boa Index]