boa.plot.title {boa}	R Documentation

Add Plot Title

Description

Adds a title to the graphics window. This function should be called after all plots have been added to the target graphics window.

Usage


boa.plot.title(text)

Arguments

text 	 Character string giving the title to be centered and displayed across the top of the graphics window.

Author(s)

Brian J. Smith

[Package boa Index]