boa.plot.trace {boa}	R Documentation

Plot Trace Histories

Description

Displays, in a single plot, the trace histories for the specified parameter(s).

Usage


boa.plot.trace(lnames, pname, annotate = boa.par("legend"))

Arguments

lnames 	 Character vector giving the name of the desired MCMC sequence in the working session list of sequences.

pname 	 Character string giving the name of the parameters to be plotted.

annotate 	 Logical value indicating that a legend be included in the plot.

Value

A logical value indicating that the plot was successfully created.

Author(s)

Brian J. Smith

See Also

boa.plot

[Package boa Index]