boa.pnames {boa}	R Documentation

Extract Parameter Names

Description

Extracts the parameter names in an MCMC sequence.

Usage


boa.pnames(link)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

Value

A character vector containing the parameter names in 'link'.

Author(s)

Brian J. Smith

See Also

boa.getparms, boa.iter

[Package boa Index]