boa.print.acf {boa}	R Documentation

Print Autocorrelation Functions

Description

Iteratively calls boa.acf() to display the lag autocorrelations for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.acf(lags = boa.par("acf.lags"))

Arguments

lags 	 Numeric vector of lags at which to estimate the autocorrelation functions.

Author(s)

Brian J. Smith

See Also

boa.acf, boa.plot, boa.plot.acf

[Package boa Index]