boa.print.cor {boa}	R Documentation

Print Correlation Matrices

Description

Iteratively computes and displays the correlation matrices for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.cor()

Author(s)

Brian J. Smith

[Package boa Index]