boa.print.geweke {boa}	R Documentation

Print Geweke Convergence Diagnostics

Description

Iteratively calls boa.geweke() to display the Geweke convergence diagnostics for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.geweke(p.first = boa.par("geweke.first"), p.last = boa.par("geweke.last"))

Arguments

p.first 	 Proportion of iterations to include in the first window.

p.last 	 Proportion of iterations to include in the last window.

Author(s)

Brian J. Smiht

See Also

boa.geweke, boa.plot, boa.plot.geweke

[Package boa Index]