boa.print.handw {boa}	R Documentation

Print Heidelberger and Welch Convergence Diagnostics

Description

Iteratively calls boa.handw() to display the Heidelberger and Welch convergence diagnostics for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.handw(error = boa.par("handw.error"), alpha = boa.par("alpha"))

Arguments

error 	 Accuracy of the posterior estimates for the monitored parameters.

alpha 	 Alpha level for the confidence in the sample mean of the retained iterations.

Author(s)

Brian J. Smith

See Also

boa.handw

[Package boa Index]