boa.print.hpd {boa}	R Documentation

Print Highest Probability Density Intervals

Description

Iteratively calls boa.hpd() to display the highest probability density (HPD) intervals for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.hpd(alpha = boa.par("alpha"))

Arguments

alpha 	 Specifies the 100*(1-alpha)% HPD intervals to be computed.

Author(s)

Brian J. Smith

See Also

boa.hpd

[Package boa Index]