boa.print.par {boa}	R Documentation

Print Global Parameters' Descriptions and Current Values

Description

Displays a description of and the current value for all global parameters.

Usage


boa.print.par(group)

Arguments

group 	 Character string specifying which parameter group to display. The groups are "Analysis", "Data", and "Plot". If omitted, information on all global parameters is displayed.

Value

The character matrix obtained from boa.pardesc() is returned, but not printed.

Author(s)

Brian J. Smith

See Also

boa.pardesc

[Package boa Index]