boa.print.randl {boa}	R Documentation

Print Raftery and Lewis Convergence Diagnostics

Description

Iteratively calls boa.randl() to display the Raftery and Lewis convergence diagnostics for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.randl(q = boa.par("randl.q"), error = boa.par("randl.error"), prob = 1 - boa.par("alpha"), delta = boa.par("randl.delta"))

Arguments

q 	 Quantile to be estimated.

error 	 Desired amount of error in estimating the specified quantile 'q'.

prob 	 Probability of attaining the desired degree of 'error'.

delta 	 Delta value used in computing the convergence diagnostic.

Author(s)

Brian J. Smith

See Also

boa.randl

[Package boa Index]