boa.print.stats {boa}	R Documentation

Print Summary Statistics

Description

Iteratively calls boa.stats() to display summary statistics for the parameters in each of the MCMC sequences in the working session list of sequences.

Usage


boa.print.stats(probs = boa.par("quantiles"), batch.size = boa.par("batch.size"))

Arguments

probs 	 Vector of probabilities at which to compute the quantiles. Values must be between 0 and 1.

batch.size 	 Number of iterations to include in each batch when computing batch means and lag-1 batch autocorrelation function.

Author(s)

Brian J. Smith

See Also

boa.stats

[Package boa Index]