boa.quit {boa}	R Documentation

Quit Session

Description

boa.quit() is called to end a BOA session. It removes the objects used to store the session lists of MCMC sequences and the global parameters. Failure to do this may result in lost or lingering data.

Usage


boa.quit()

Author(s)

Brian J. Smith

See Also

boa.save, boa.init

[Package boa Index]