boa.randl {boa}	R Documentation

Raftery and Lewis Convergence Diagnostics

Description

Computes the Raftery and Lewis convergence diagnostics for the parameters in an MCMC sequence.

Usage


boa.randl(link, q, error, prob, delta)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

q 	 Quantile to be estimated.

error 	 Desired amount of error in estimating the specified quantile 'q'.

prob 	 Probability of attaining the desired degree of error - 'error'.

delta 	 Delta value used in computing the convergence diagnostics.

Value

A matrix whose columns and rows are the Raftery and Lewis convergence diagnostics (i.e. thin, burn-in, total, lower bound, and dependence factor) and the monitored parameters, respectively.

Author(s)

Brian J. Smith, Nicky Best, Kate Cowles

References

Raftery, A. L. and Lewis, S. (1992a). Comment: One long run with diagnostics: Implementation strategies for Markov chain Monte Carlo. Statistical Science, 7, 493-7.
Raftery, A. L. and Lewis, S. (1992b). How many iterations in the Gibbs sampler? In Bayesian Statistics 4, (ed. J. M. Bernardo, J. O. Berger, A. P. Dawid, and A. F. M. Smith), pp. 763-74. Oxford University Press.

See Also

boa.print.randl

[Package boa Index]