boa.save {boa}	R Documentation

Save Session

Description

Save the current state of the session lists of MCMC sequences and the global parameters to a database object.

Usage


boa.save(name, envir = globalenv())

Arguments

name 	 Character string giving the name of the object to which the current session should be saved.

envir 	 The 'environment' to which the object should be saved. For more information, consult the help documentation in R on the assign() function.

Value

A logical value indicating that the session was successfully saved to the specified object.

Author(s)

Brian J. Smith

[Package boa Index]