boa.sortparms {boa}	R Documentation

Sort by Parameter Names

Description

Sorts the parameters of an MCMC sequence alphabetically by parameter names.

Usage


boa.sortparms(link)

Arguments

link 	 Matrix whose columns and rows contain the monitored parameters and the MCMC iterations, respectively. The iteration numbers and parameter names must be assigned to dimnames(link).

Value

A copy of 'link' with the columns sorted alphabetically by parameter names.

Author(s)

Brian J. Smith

See Also

boa.sortiter

[Package boa Index]