boa.transform {boa}	R Documentation

Transform Variables

Description

Applies a log or logit transformation to map bounded variables to the entire real line.

Usage


boa.transform(x, support, inv = FALSE)

Arguments

x 	 Numeric object to be transformed.

support 	 Vector giving the lower and upper limits of the support for the untransformed data. Unbounded lower or upper limits should be specified as -Inf or Inf, respectively.

inv 	 Logical value indicating that the inverse transformation should be applied.

Value

Transformed or inverse-transformed analogue of 'x'.

Author(s)

Brian J. Smith

[Package boa Index]