line {boa}	R Documentation

BUGS Line Example

Description

The Line example involves a linear regression analysis of the data points (1,1), (2,3), (3,3), (4,3), and (5,5). The proposed Bayesian model is

y[i] ~ N(mu[i], tau)

mu[i] = alpha + beta * (x[i] - mean(x[]))

with the following priors:

alpha ~ N(0, 0.0001)

beta ~ N(0, 0.0001)

tau ~ Gamma(0.001, 0.001)

Two parallel chains from the MCMC sampler were generated and the output saved in the S data frames line1 and line2.

Usage

data(line)

Format

The data frames line1 and line2 each containing 200 observations.

References

Spiegelhalter, D., Thomas, A. and Best, N., (2000). WinBugs Version 1.4 User Manual.

[Package boa Index]